frontend/requirements.txt
streamlit
requests
matplotlib
pandas
python-dotenv
groq
